HidingHook: Introduce Module Hiding Method
Some applications will check for Ad-Blockers through some system API, so we decide to introduce module hiding. It maybe a no-op because anyway applications can use reflection to restrict an Xposed module, but it is better than none.

The method is taken from Xposed module WechatLuckyMoney, and modified by me.